% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{convert_data_to_numeric}
\alias{convert_data_to_numeric}
\title{Convert non-numeric to numeric for XGB}
\usage{
convert_data_to_numeric(x)
}
\description{
Convert non-numeric to numeric for XGB
}
\keyword{internal}
