% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{ds_xgb}
\alias{ds_xgb}
\title{XGBoost}
\usage{
ds_xgb(space, data, ...)
}
\arguments{
\item{space}{(\code{character(1)}) \cr
Name of the V-Dem indicator for a democratic space}

\item{data}{(\code{\link[base:data.frame]{base::data.frame()}})
A data frame or similar object}

\item{...}{document other arguments}
}
\description{
Demonstrates the interface. Otherwise only difference is that it internally
normalizes input data before fitting and predicting.
}
\examples{
data("states")

mdl <- ds_xgb("v2x_veracc_osp", states)
preds <- predict(mdl, new_data = states)

}
