% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_option}
\alias{get_option}
\title{Option accessor}
\usage{
get_option(x)
}
\arguments{
\item{x}{Option name, either short like "version" or with the prefix like
"demspaces.version".}
}
\value{
Option value, with correct type.
}
\description{
Get DemSpaces internal package options, with additional safety checks. I.e.
I don't want to build the merge data with a NULL value for start_year.
}
