% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_forecasts}
\alias{read_forecasts}
\title{Read the forecasts}
\usage{
read_forecasts(version = "latest")
}
\arguments{
\item{version}{Data version, e.g. "v13"; default is the latest available.}
}
\description{
Read the forecast data.
}
\details{
This function will preferentially look for data in
\code{modelrunner/output} so that during updates the latest available data is
read. As a backup it will check in \code{archive}.
}
