% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_merge_data}
\alias{read_merge_data}
\title{Read the merge data}
\usage{
read_merge_data(version = "latest")
}
\arguments{
\item{version}{Data version, e.g. "v13"; default is the latest available.}
}
\description{
Read the merge data.
}
\details{
This function will preferentially look for data in
\code{create-data/output} so that during updates the latest available data is
read. As a backup it will check in \code{archive/data}.
}
