% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{xgb}
\alias{xgb}
\title{XGBoost}
\usage{
xgb(x, y)
}
\arguments{
\item{x}{Data frame with features.}

\item{y}{Binary vector indicating outcome event.}
}
\description{
Standardized interface for ...
}
\examples{
credit_data <- recipes::credit_data

mdl <- xgb(credit_data[, setdiff(colnames(credit_data), "Status")],
                    credit_data$Status)

}
